Add the file deletion logic to library.go.
The delete-file command is going to need the added methods to function.

For now, the use case of deleting a book after deleting it's last file is not yet implemented.
That means that books with no files might exist in the database. This is, of course, going to be fixed.

The Library.IsLastFile method added in this commit is going to be used by the delete-file command to display a warning before deleting a file from a book that has no other files (as requested by Tyler).
It will probably also be needed to detect when to trigger the book deletion behaviour when we implement it.