Add the delete-file command.
The delete-file command deletes a file by its ID or filename (if the -f flag is used).
If the deleted file is the only file in a book, a warning is displayed instead. To delete it anyway, the -y flag needs to be passed.

For now deleting such a file leaves an empty book (this is going to be fixed).